+++ widgets_demo/lib/components/charts/presenters/container.dart
@@ -77,5 +77,51 @@ WidgetbookComponent containerChart(BuildContext context) {
       },
       "markdown/charts_container_indicateurs.md",
     ),
+    usercaseWithMarkdown(
+      "Second bouton",
+      (context) {
+        return Container(
+          color: SepteoColors.blue.shade50,
+          child: Center(
+            child: SizedBox(
+              width: 350,
+              child: Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: Center(
+                  child: ChartContainer(
+                    buttonText: "Export des bulletins (17)",
+                    buttonIcon: Icon(Icons.file_download_outlined),
+                    onButtonPressed: () {},
+                    secondaryButtonIcon: Icon(Icons.undo),
+                    onSecondaryButtonPressed: () {},
+                    secondaryButtonText: "Annuler",
+                    title: "Etat des bulletins",
+                    child: ChartPie(items: [
+                      ChartPieItem(
+                          value: 50,
+                          title: 'Calculé',
+                          color: SepteoColors.blue.shade600),
+                      ChartPieItem(
+                          value: 25,
+                          title: 'Envoyé',
+                          color: SepteoColors.green.shade600),
+                      ChartPieItem(
+                          value: 6,
+                          title: 'Non calculé',
+                          color: SepteoColors.red.shade600),
+                      ChartPieItem(
+                          value: 19,
+                          title: 'En cours',
+                          color: SepteoColors.orange.shade600),
+                    ]),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        );
+      },
+      "markdown/charts_container_pie_second_button.md",
+    ),
   ]);
 }