+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 0.6.0
+## 0.6.0+1
 * Replaced local theme with septeo theme using septeo_design_system
+* Add Secondary Button on ChartContainer
+* Add tri state checkbox
+* Add CheckBoxWithText  widget
 
 ## 0.5.0+3
 * Add Charts